+++ packages/terra-graphs-docs/src/terra-dev-site/doc/example/Graphs/Timeline/Criticality/TimelineCriticality.jsx
@@ -1,0 +1,22 @@
+import React from 'react';
+import Timeline from '@cerner/terra-graphs/lib/components/Timeline/Timeline';
+import utils from '@cerner/carbon-graphs/lib/js/helpers/utils';
+import '@cerner/terra-graphs-docs/lib/Css/ExampleGraphContainer.module.scss';
+import getTimelineConfig from '@cerner/terra-graphs-docs/lib/example-datasets/graphConfigObjects/Timeline/timelineConfigObject';
+import data from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/Timeline/defaultDataset.js';
+
+/*
+Please refer to the documentation below to see the graphConfig and data objects
+*/
+const graphConfig = utils.deepClone(getTimelineConfig('#simpleTimelineGraph'));
+const dataset = utils.deepClone(data);
+dataset[0].values[0].isCritical = true;
+dataset[0].values[1].isCritical = true;
+dataset[0].values[3].isCritical = true;
+
+export default () => (
+  <React.Fragment>
+    <div id="tooltip" className="initial-tooltip" />
+    <Timeline graphID="simpleTimelineGraph" graphConfig={graphConfig} dataset={dataset} />
+  </React.Fragment>
+);