+++ packages/terra-graphs-docs/src/terra-dev-site/test/graphs/Timeline/BasicTimeline/AxisWithNoTickValues.test.jsx
@@ -5,12 +5,11 @@ import '@cerner/terra-graphs-docs/lib/Css/ExampleGraphContainer.module.scss';
 import getTimelineConfig from '@cerner/terra-graphs-docs/lib/example-datasets/graphConfigObjects/Timeline/timelineConfigObject';
 import data from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/Timeline/defaultDataset.js';
 
-/*
-Please refer to the documentation below to see the graphConfig and data objects
-*/
 const graphConfig = utils.deepClone(getTimelineConfig('#simpleTimelineGraph'));
-const dataset = utils.deepClone(data[0]);
-
+const dataset = utils.deepClone(data);
+graphConfig.axis.x.ticks = {
+  format: '',
+};
 export default () => (
   <React.Fragment>
     <div id="tooltip" className="initial-tooltip" />